fix: call stack overflow in webtransport transport
Fixes bug in webtransport where the `.abort` method on the stream class
was calling `stream.abort` and leading to a call stack overflow.
