+++ packages/transport-webtransport/test/browser.ts
@@ -5,7 +5,7 @@ import { noise } from '@chainsafe/libp2p-noise'
 import { multiaddr } from '@multiformats/multiaddr'
 import { expect } from 'aegir/chai'
 import { createLibp2p } from 'libp2p'
-import { webTransport, isSubset } from '../src/index'
+import { webTransport } from '../src/index.js'
 
 declare global {
   interface Window {
@@ -124,7 +124,7 @@ describe('libp2p-webtransport', () => {
     await node.stop()
   })
 
-  it('Closes writes of streams after they have sunk a source', async () => {
+  it('closes writes of streams after they have sunk a source', async () => {
     // This is the behavior of stream muxers: (see mplex, yamux and compliance tests: https://github.com/libp2p/js-libp2p-interfaces/blob/master/packages/interface-stream-muxer-compliance-tests/src/close-test.ts)
     if (process.env.serverAddr == null) {
       throw new Error('serverAddr not found')
@@ -167,23 +167,5 @@ describe('libp2p-webtransport', () => {
     expect(stream.timeline.close).to.be.greaterThan(0)
 
     await node.stop()
-  })
-})
-
-describe('test helpers', () => {
-  it('correctly checks subsets', () => {
-    const testCases = [
-      { a: [[1, 2, 3]], b: [[4, 5, 6]], isSubset: false },
-      { a: [[1, 2, 3], [4, 5, 6]], b: [[1, 2, 3]], isSubset: true },
-      { a: [[1, 2, 3], [4, 5, 6]], b: [], isSubset: true },
-      { a: [], b: [[1, 2, 3]], isSubset: false },
-      { a: [], b: [], isSubset: true },
-      { a: [[1, 2, 3]], b: [[1, 2, 3], [4, 5, 6]], isSubset: false },
-      { a: [[1, 2, 3]], b: [[1, 2]], isSubset: false }
-    ]
-
-    for (const tc of testCases) {
-      expect(isSubset(tc.a.map(b => new Uint8Array(b)), tc.b.map(b => new Uint8Array(b)))).to.equal(tc.isSubset)
-    }
   })
 })
